Move tabs variable into status_header.js.
